Add a HEAD option for setting maximum number of time series
This option allows us to set a limit for the number of series that can be stored on HEAD.
When sample append would require a new time series to be created, but we already have MaxSeries number of series stored on HEAD then append will return an error.
Appends for samples that would append values to already stores series will be accepted.

This effectively implements load shedding for Prometheus and allows to set a capacity limit for the number of time series it will accept.
There are already ways to set per scrape pool limits, but they are difficult to use from an operator point of view, since they work on each individual pool level and so the total consumed capacity depends on the number of targets in each pool and the actual number of metrics being exported. With enviroment complex or dynamic enough this can be tricky to manage.
Having a single number that forces the number of metrics is much easier to work with.

Since all series are kept in-memory until a block is written, after which series with no new samples might get garbage collected, not allowing HEAD to create new series when we're at capacity is the best way to ensure predictable memory usage. Samples being appended to already existing series are less likely to cause memory usage problems, so they are allowed to continue.

Signed-off-by: Łukasz Mierzwa <l.mierzwa@gmail.com>