Don't mark targets as down when out of capacity
When Prometheus runs out of capacity it might not ingest all samples from targets, so an error is returned on append.
But this could lead to a flood of confusin alert saying that lots of targets are down, simply because we had to drop a sample from each of them.
To avoid this let's (for now) ignore capacity errors when checking target health.

Signed-off-by: Łukasz Mierzwa <l.mierzwa@gmail.com>