Use a linked list for memSeries.headChunk
Currently memSeries holds a single head chunk in-memory and a slice of mmapped chunks.
When append is called memSeries might decide that it needs to create a new head chunk and append it there,
for that it will first mmap existing head chunk and create a new one, only after that happens it will continue appending our sample.
Since appending samples uses write lock on memSeries no other read or write can happen until any append is completed.
So when we have an append that must create a new head chunk the whole memSeries is blocked until mmapping of existing head chunk finishes.
Mmapping itself uses a lock as it needs to be serialised, which means that the more chunks to mmap we have the longer each chunk might wait
for it to be mmapped.
If there's enough chunks that require mmapping some memSeries will be locked for more than query timeout which will fail queries and appends.

To avoid this we need to remove mmapping from append path, since mmapping is blocking.
But this means that when we cut a new head chunk we need to keep the old one around, so we can mmap it later.
This change makes memSeries.headChunk a linked list, memSeries.headChunk still points to the 'open' head chunk that receives new samples,
while older, yet to be mmapped, chunks are linked to it.
Mmapping is done on a schedule by iterating all memSeries one by one. Thanks to this when control when mmapping is done, since we trigger
it manually, which reduces the risk that it will have to compete for mmap locks with other chunks.

Fixes #10377

Signed-off-by: Łukasz Mierzwa <l.mierzwa@gmail.com>