Reclaim unused memory in scrapeCache
scrapeCache is a heavy user of Go maps.
Go maps are fast but they never shrink, so if we add a huge number of elements to a map, then remove half of them, our map will still use memory needed to store peak number of elements.
Over a long runtime period, if there's enough churn and spikes, we might end up keeping over-allocated maps around.
This change tries to address this by re-creating maps if elements were deleted from them.
Because it's no possible to get the capacity of a map in Go there isn't really a good way of testing this change.

Signed-off-by: Łukasz Mierzwa <l.mierzwa@gmail.com>