Shard MemPostings
Currently all access to MemPostings is via a single mutex, which in effect serializes all Add() calls.
This can cause MemPostings.Add() to get blocked for a long time in some cases, mostly when:
- Prometheus was just started
- There's a lot of scrape targets with plenty of metrics
- There are indentical label key/value pairs attached to a lot of time series
- There is no previous WAL files

With such conditions all initial scrapes will happens withing the first scrape interval and will
have to append entries to MemPostings causing high lock contention on that single mutex.
Presence of 'common' labels that are attached to a lot of time series means that there will be
a few entries in MemPostings where the slice of SeriesRef will be very big and that will trigger
a lot of sorting calls - due to the fact that SeriesRef are oredered in the slices stored by MemPostings.
This need to ensure order will make MemPostings.Add() calls slow and cpu intensive.
All of that is problematic but usually most of it happens during WAL replay, so it's only visible
on scraped data if there was no WAL to replay.

There is https://github.com/prometheus/prometheus/pull/13632 that tries to speed up the code
responsible for sorting SeriesRef in MemPostings slices.

This change is to also shard MemPostings data structure so instead of a single lock we have
a number of locks. Sharding happens on the label name so reads & writes on different label names
might end up on a different shard and so allow a level of concurrency.
This is similiar to how stripeSeries in tsdb/head.go works.

Signed-off-by: Łukasz Mierzwa <l.mierzwa@gmail.com>